remove root `./test` directory.
