+++ packages/frint-test-utils/register.js
@@ -2,3 +2,5 @@ var lib = require('./lib');
 
 global.resetDOM = lib.resetDOM;
 global.resetDOM();
+
+lib.takeOverConsole(console);
